refactor(bool fn trait): moved expression's iterator blocks back to src/expression
